Responsive Layout
Made the main page more responsive. By default, the layout consists of two
columns, a content bar and a contact bar. When the screen reaches tablet width,
it resizes so that the contact bar stacks on top of the content. Made certain
alterations to the layout as well, including moving the header into includes.